Move internal ID2D1DrawTransformMapperInterop type
