+++ src/ComputeSharp.D2D1/Shaders/Interop/Effects/TransformMappers/D2D1DrawTransformMapperImpl.ID2D1DrawTransformMapper.cs
@@ -3,7 +3,6 @@ using System.Drawing;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using ComputeSharp.D2D1.Extensions;
-using ComputeSharp.D2D1.Interop.Effects;
 using ComputeSharp.Win32;
 
 namespace ComputeSharp.D2D1.Shaders.Interop.Effects.TransformMappers;
