+++ src/ComputeSharp.D2D1/Shaders/Interop/Effects/TransformMappers/D2D1DrawTransformMapperImpl.cs
@@ -4,7 +4,6 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Threading;
 using ComputeSharp.D2D1.Interop;
-using ComputeSharp.D2D1.Interop.Effects;
 using ComputeSharp.Win32;
 
 namespace ComputeSharp.D2D1.Shaders.Interop.Effects.TransformMappers;
